backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
sentence-transformers>=2.3.0
lancedb>=0.5.0
numpy>=1.26.0
umap-learn>=0.5.5
orjson>=3.9.10
python-multipart>=0.0.9
torch>=2.1.0
pyarrow>=15.0.0
xxhash>=3.4.1
scikit-learn>=1.4.0
joblib>=1.3.0
tqdm>=4.66.0
pandas>=2.0.0
pylance>=0.10.0
plotly>=5.18.0
rich>=13.0.0
